+++ namesys/resolve/resolve.go
@@ -1,17 +1,20 @@
-package core
+package resolve
 
 import (
 	"context"
 	"errors"
 	"strings"
 
-	namesys "github.com/ipfs/go-ipfs/namesys"
-
-	ipld "github.com/ipfs/go-ipld-format"
+	"github.com/ipfs/go-ipld-format"
+	log2 "github.com/ipfs/go-log"
 	logging "github.com/ipfs/go-log"
-	path "github.com/ipfs/go-path"
-	resolver "github.com/ipfs/go-path/resolver"
+	"github.com/ipfs/go-path"
+	"github.com/ipfs/go-path/resolver"
+
+	"github.com/ipfs/go-ipfs/namesys"
 )
+
+var log = logging.Logger("nsresolv")
 
 // ErrNoNamesys is an explicit error for when an IPFS node doesn't
 // (yet) have a name system
@@ -27,34 +30,34 @@ func ResolveIPNS(ctx context.Context, nsys namesys.NameSystem, p path.Path) (pat
 
 		// TODO(cryptix): we should be able to query the local cache for the path
 		if nsys == nil {
-			evt.Append(logging.LoggableMap{"error": ErrNoNamesys.Error()})
+			evt.Append(log2.LoggableMap{"error": ErrNoNamesys.Error()})
 			return "", ErrNoNamesys
 		}
 
 		seg := p.Segments()
 
 		if len(seg) < 2 || seg[1] == "" { // just "/<protocol/>" without further segments
-			evt.Append(logging.LoggableMap{"error": path.ErrNoComponents.Error()})
+			evt.Append(log2.LoggableMap{"error": path.ErrNoComponents.Error()})
 			return "", path.ErrNoComponents
 		}
 
 		extensions := seg[2:]
 		resolvable, err := path.FromSegments("/", seg[0], seg[1])
 		if err != nil {
-			evt.Append(logging.LoggableMap{"error": err.Error()})
+			evt.Append(log2.LoggableMap{"error": err.Error()})
 			return "", err
 		}
 
 		respath, err := nsys.Resolve(ctx, resolvable.String())
 		if err != nil {
-			evt.Append(logging.LoggableMap{"error": err.Error()})
+			evt.Append(log2.LoggableMap{"error": err.Error()})
 			return "", err
 		}
 
 		segments := append(respath.Segments(), extensions...)
 		p, err = path.FromSegments("/", segments...)
 		if err != nil {
-			evt.Append(logging.LoggableMap{"error": err.Error()})
+			evt.Append(log2.LoggableMap{"error": err.Error()})
 			return "", err
 		}
 	}
@@ -64,7 +67,7 @@ func ResolveIPNS(ctx context.Context, nsys namesys.NameSystem, p path.Path) (pat
 // Resolve resolves the given path by parsing out protocol-specific
 // entries (e.g. /ipns/<node-key>) and then going through the /ipfs/
 // entries and returning the final node.
-func Resolve(ctx context.Context, nsys namesys.NameSystem, r *resolver.Resolver, p path.Path) (ipld.Node, error) {
+func Resolve(ctx context.Context, nsys namesys.NameSystem, r *resolver.Resolver, p path.Path) (format.Node, error) {
 	p, err := ResolveIPNS(ctx, nsys, p)
 	if err != nil {
 		return nil, err
